refactor(schema): remove dependency_graph table
Phase 3 of dependency consolidation (Issue #37).

Removed all code related to the unused dependency_graph table:

Schema changes:
- Removed dependency_graph table definition (lines 234-294)
- Removed 7 associated indexes

Code removal from queries.ts:
- storeDependenciesInternal()
- storeDependencies()
- queryDependentsRaw()
- queryDependenciesRaw()
- storeDependenciesLocal()
- queryDependentsLocal alias
- queryDependenciesLocal alias
- Dependency extraction from runIndexingWorkflow()
- dependenciesExtracted from return type

Files deleted:
- app/src/indexer/dependency-extractor.ts (353 lines)

Other changes:
- Moved DependencyEdge type to circular-detector.ts (only remaining consumer)
- Updated all test files to remove dependency graph test cases
- Updated MCP tools to remove dependenciesExtracted from results

Total reduction: ~650 lines of code
